+++ CHANGELOG.md
@@ -2,10 +2,18 @@
 
 ## UNRELEASED
 
+-   `Card`
+
+    -   Fixed missing `<Card.Footer fit radius shape>` support, to match `<Figure>` Property API.
+
 -   `keybind`
 
     -   Fixed `make_keybind_shortcut` missing as a package export.
     -   Updated `make_keybind_shortcut` action to allow override of preset options aside from `IKeybindOptions.on_bind`.
+
+-   `Tile`
+
+    -   Fixed missing `<Tile.Footer fit radius shape>` support, to match `<Figure>` Property API.
 
 ## v0.6.3 - 2022/03/11